Pet Classes with Interface
