+++ src/edu/htc/pets/Audible.java
@@ -1,0 +1,9 @@
+package edu.htc.pets;
+
+/**
+ * Created by cheey on 2/29/2016.
+ */
+public interface Audible {
+
+    public void makeSound();
+}
