+++ src/edu/htc/pets/Pet.java
@@ -1,23 +1,14 @@
-/**
- * Created by cheey on 2/22/2016.
- */
 package edu.htc.pets;
-public abstract class Pet //ABSTRACT CLASS
-{
 
-    private String name;
+/**
+ * Created by Student on 2/22/2016.
+ */
+public abstract class  Pet {
+
     private int age;
+    private String name;
 
-    public Pet(String name)
-    {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
+    public Pet(String name) {
         this.name = name;
     }
 
@@ -29,9 +20,34 @@ public abstract class Pet //ABSTRACT CLASS
         this.age = age;
     }
 
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
     public String toString() {
         return "Name: " + getName();
     }
 
     public abstract double getHumanAge();
+
+
+    public static void main(String[] blah){
+        Pet pet = new Pet("Name") {
+
+            public double getHumanAge(){
+                return 0;
+            }
+
+        };
+
+        String name = pet.getName();
+        if ( ! name.equals("Name")){
+            System.out.println("Something went wrong in my constructor");
+        }
+
+    }
 }
